requirements.txt
Flask
python-dotenv
flask-restful
flask_sqlalchemy<2.0
pika
rabbitmq